+++ config.server.ts
@@ -354,6 +354,7 @@ export default defineConfig<Config>(
                             checkout_step: true,
                             view_search_suggestion: true,
                             click_search_suggestion: true,
+                            commerce_agent_engagement: true,
                         },
                     },
                     dataCloud: {
@@ -375,6 +376,7 @@ export default defineConfig<Config>(
                             checkout_step: true,
                             view_search_suggestion: true,
                             click_search_suggestion: true,
+                            commerce_agent_engagement: true,
                         },
                     },
                     activeData: {
@@ -395,6 +397,7 @@ export default defineConfig<Config>(
                             checkout_step: false,
                             view_search_suggestion: false,
                             click_search_suggestion: false,
+                            commerce_agent_engagement: true,
                         },
                     },
                 },