+++ src/test-utils/config.ts
@@ -237,6 +237,7 @@ export const mockBuildConfig: Config = {
                         checkout_step: true,
                         view_search_suggestion: true,
                         click_search_suggestion: true,
+                        commerce_agent_engagement: true,
                     },
                 },
                 dataCloud: {
@@ -258,6 +259,7 @@ export const mockBuildConfig: Config = {
                         checkout_step: true,
                         view_search_suggestion: true,
                         click_search_suggestion: true,
+                        commerce_agent_engagement: true,
                     },
                 },
                 activeData: {
@@ -280,6 +282,7 @@ export const mockBuildConfig: Config = {
                         checkout_step: false,
                         view_search_suggestion: false,
                         click_search_suggestion: false,
+                        commerce_agent_engagement: true,
                     },
                 },
             },